refactor: removes onchain publishing
