+++ .env-example
@@ -1,16 +1,14 @@
 ### FOR TESTING ###
 
-WALLET_KEY=
-
 AUDIT_ENDPOINT="http://localhost:3000/api/audit/"
+
+USER_ENDPOINT = "http://localhost:3000/api/user/auditor"
 
 PROJECT_ENDPOINT="http://localhost:3000/api/project/"
 
 FORWARDER_ENDPOINT="http://localhost:3000/api/audit/forwarder/"
 
 WEB3_STORAGE_API_ENDPOINT="http://localhost:3000/api/file/"
-
-TB_CORE_ADDRESS=
 
 MASTER_KEY=MY_MASTER_KEY
