+++ Cargo.toml
@@ -59,8 +59,7 @@ clap = { version = "4.1.6", features = ["derive", "unstable-doc"] }
 color-eyre = "0.6"
 dirs = "5.0.0"
 dotenv = "0.15.0"
-ethers = { version = "2.0.1", features = ["eip712"] }
-ethers-core = { version = "2.0.1", features = ["eip712"] }
+ethers-core = "2.0.1"
 eyre = "0.6.8"
 futures = "0.3"
 futures-util = { version = "0.3.17", default-features = false, features = [
