+++ docker-compose.yml
@@ -1,41 +1,40 @@
 version: "3.9"
 services:
-  db:
-    image: mysql:latest
-    container_name: tb-db
-    restart: always
-    environment:
-      MYSQL_ALLOW_EMPTY_PASSWORD: true
-      MYSQL_ROOT_PASSWORD: root
-      MYSQL_DATABASE: local
-      MYSQL_USER: user
-      MYSQL_PASSWORD: pass
-    ports:
-      - "3306:3306"
-    healthcheck:
-      test: mysqladmin ping -h 127.0.0.1 -u $$MYSQL_USER --password=$$MYSQL_PASSWORD
-      interval: 3s
-      timeout: 20s
-      retries: 30
+    db:
+        image: mysql:latest
+        container_name: tb-db
+        restart: always
+        environment:
+            MYSQL_ALLOW_EMPTY_PASSWORD: true
+            MYSQL_ROOT_PASSWORD: root
+            MYSQL_DATABASE: local
+            MYSQL_USER: user
+            MYSQL_PASSWORD: pass
+        ports:
+            - "3306:3306"
+        healthcheck:
+            test: mysqladmin ping -h 127.0.0.1 -u $$MYSQL_USER --password=$$MYSQL_PASSWORD
+            interval: 3s
+            timeout: 20s
+            retries: 30
 
-  app:
-    image: teamauth/webapp:latest-staging
-    container_name: tb-webapp
-    depends_on:
-      db:
-       condition: service_healthy
-    ports:
-      - "3000:3000"
-    env_file:
-      - .env.development.local
-  pdf:
-    image: tb-pdf
-    container_name: tb-pdf
-    depends_on:
-      db:
-        condition: service_healthy
-      app:
-        condition: service_started
-    ports:
-      - "3005:3005"
-
+    app:
+        image: teamauth/webapp:latest
+        container_name: tb-webapp
+        depends_on:
+            db:
+                condition: service_healthy
+        ports:
+            - "3000:3000"
+        env_file:
+            - .env.development.local
+    pdf:
+        image: tb-pdf
+        container_name: tb-pdf
+        depends_on:
+            db:
+                condition: service_healthy
+            app:
+                condition: service_started
+        ports:
+            - "3005:3005"
