+++ src/cmd/init.rs
@@ -9,9 +9,6 @@ use clap::Parser;
 #[allow(clippy::module_name_repetitions)]
 #[derive(Debug, Clone, Parser)]
 pub struct InitArgs {
-    #[clap(short, long, help = "Private key to add to .env")]
-    private_key: Option<String>,
-
     #[clap(
         short,
         long,
@@ -25,7 +22,6 @@ impl Cmd for InitArgs {
     fn run(self) -> eyre::Result<()> {
         let home_dir = dirs::home_dir().wrap_err("Could not find home directory")?;
 
-        let private_key = self.private_key.unwrap_or_default();
         let api_key = self.api_key.unwrap_or_default();
 
         // Create the path to the .trustblock directory
@@ -46,7 +42,7 @@ impl Cmd for InitArgs {
 
         let mut env_file = File::create(&env_path)?;
 
-        let env_data = format!("WALLET_KEY={private_key}\nAPI_KEY={api_key}");
+        let env_data = format!("API_KEY={api_key}");
 
         env_file.write_all(env_data.as_bytes())?;
