+++ src/constants.rs
@@ -1,20 +1,6 @@
-pub const GOERLI_RPC_URL: &str =
-    "https://proportionate-cool-feather.ethereum-goerli.quiknode.pro/8dd07d504269d8367298f9034bcd7d954c262977/";
-
-pub const MUMBAI_RPC_URL: &str =
-    "https://few-empty-hill.matic-testnet.quiknode.pro/b39218cd6c5dbf671082e8cfb4cbcf046d3e56cc/";
-
-pub const BNB_TESTNET_RPC_URL: &str =
-    "https://fittest-dry-sunset.bsc-testnet.quiknode.pro/75d26b70af2314bce4c425561a109470ce229e7a/";
-
-pub const FUJI_RPC_URL: &str =
-    "https://shy-cool-bird.avalanche-testnet.quiknode.pro/9d3bf2d2581e422a028007865d29e741bb94029f/ext/bc/C/rpc";
-
 pub const AUDIT_ENDPOINT: &str = "https://beta.trustblock.run/api/audit/";
 
 pub const PROJECT_SLUG_ENDPOINT: &str = "https://beta.trustblock.run/api/project/";
-
-pub const FORWARDER_ENDPOINT: &str = "https://beta.trustblock.run/api/audit/forwarder/";
 
 pub const WEB3_STORAGE_API_ENDPOINT: &str = "https://beta.trustblock.run/api/file/";
 
@@ -27,16 +13,6 @@ pub const GITHUB_LATEST_RELEASE: &str =
 
 pub const WEB3_STORAGE_ENDPOINT: &str = ".ipfs.w3s.link";
 
-pub const FORWARDER_GENERIC_PARAMETERS: &str =
-    "address from,address to,uint256 value,uint256 gas,uint256 nonce,bytes data,uint256 validUntilTime";
-
-pub const FORWARDER_DOMAIN_NAME: &str = "Trustblock Forwarder";
-pub const FORWARDER_DOMAIN_VERSION: &str = "1";
-
 pub const TRUSTBLOCK_API_KEY_HEADER: &str = "x-trustblock-api-key";
 
 pub const CLI_PATH: &str = ".trustblock";
-
-pub const TB_CORE_ADDRESS: &str = "0x3299d8FD5d3689Dd1f183cFBCeC986FAC63c8B8D";
-
-pub const FORWARDER_ADDRESS: &str = "0x0dD4D7f93Bcc77dF99d516660D33697aa695BA46";
