--- src/types/trustblock_contracts.rs
@@ -1,7 +1,0 @@
-use ethers::contract::abigen;
-
-abigen!(
-    TrustblockForwarder,
-    "src/data/Forwarder.json",
-    event_derives(serde::Deserialize, serde::Serialize)
-);
