+++ tests/test-data/fixtures/env_file_no_values.stdout
@@ -1,2 +1,1 @@
-WALLET_KEY=
 API_KEY=
