+++ tests/test-data/fixtures/env_file_with_values.stdout
@@ -1,2 +1,1 @@
-WALLET_KEY=some_private_key
 API_KEY=some_api_key
