+++ tests/test-data/fixtures/publish_db_report_exists.stdout
@@ -4,5 +4,3 @@
 )+\[\+\] Uploading is done
 
 Audit already published to DB!
-
-Audit published successfully
